Create ExpenseController and ExpenseControllerTest

+++ src/main/java/com/myMoneyTracker/controller/IncomeController.java
@@ -1,21 +1,30 @@
 package com.myMoneyTracker.controller;
 
-import com.myMoneyTracker.converter.IncomeConverter;
-import com.myMoneyTracker.dao.IncomeDao;
-import com.myMoneyTracker.dto.income.IncomeDTO;
-import com.myMoneyTracker.model.income.Income;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.EmptyResultDataAccessException;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import javax.validation.ConstraintViolationException;
-import javax.validation.Valid;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.logging.Level;
 import java.util.logging.Logger;
+
+import javax.validation.ConstraintViolationException;
+import javax.validation.Valid;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.EmptyResultDataAccessException;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RestController;
+
+import com.myMoneyTracker.converter.IncomeConverter;
+import com.myMoneyTracker.dao.AppUserDao;
+import com.myMoneyTracker.dao.IncomeDao;
+import com.myMoneyTracker.dto.income.IncomeDTO;
+import com.myMoneyTracker.model.income.Income;
+import com.myMoneyTracker.model.user.AppUser;
+import com.myMoneyTracker.util.ControllerUtil;
 
 /**
  * @author Floryn
@@ -30,6 +39,9 @@ public class IncomeController {
     IncomeDao incomeDao;
     
     @Autowired
+    private AppUserDao appUserDao;
+    
+    @Autowired
     IncomeConverter incomeConverter;
     
     private static final Logger log = Logger.getLogger(AppUserController.class.getName());
@@ -38,6 +50,9 @@ public class IncomeController {
     public ResponseEntity<?> createIncome(@RequestBody @Valid Income income) {
     
         try {
+            String loggedUsername = ControllerUtil.getCurrentLoggedUsername();
+            AppUser user = appUserDao.findByUsername(loggedUsername);
+            income.setUser(user);
             Income createdIncome = incomeDao.saveAndFlush(income);
             return new ResponseEntity<IncomeDTO>(incomeConverter.convertTo(createdIncome), HttpStatus.OK);
         } catch (ConstraintViolationException e) {
@@ -49,7 +64,8 @@ public class IncomeController {
     @RequestMapping(value = "/find_all", method = RequestMethod.GET)
     public ResponseEntity<List<IncomeDTO>> listAllIncomes() {
     
-        List<Income> incomes = incomeDao.findAll();
+        String loggedUsername = ControllerUtil.getCurrentLoggedUsername();
+        List<Income> incomes = incomeDao.findByUsername(loggedUsername);
         if (incomes.isEmpty()) {
             return new ResponseEntity<List<IncomeDTO>>(HttpStatus.NO_CONTENT);
         }
@@ -59,31 +75,24 @@ public class IncomeController {
     @RequestMapping(value = "/find/{id}", method = RequestMethod.GET)
     public ResponseEntity<?> findIncome(@PathVariable("id") Long id) {
     
+        String loggedUsername = ControllerUtil.getCurrentLoggedUsername();
         Income income = incomeDao.findOne(id);
-        if (income == null) {
+        if (income == null || !(loggedUsername.equals(income.getUser().getUsername()))) {
             return new ResponseEntity<String>("Income not found", HttpStatus.NOT_FOUND);
         }
         return new ResponseEntity<IncomeDTO>(incomeConverter.convertTo(income), HttpStatus.OK);
     }
     
-    @RequestMapping(value = "/find/user/{user_id}", method = RequestMethod.GET)
-    public ResponseEntity<?> findByUserId(@PathVariable("user_id") Long id) {
-    
-        List<Income> incomeList = incomeDao.findByUserId(id);
-        if (incomeList == null || incomeList.isEmpty()) {
-            return new ResponseEntity<String>("Income not found", HttpStatus.NOT_FOUND);
-        }
-        return new ResponseEntity<List<IncomeDTO>>(createIncomeDTOs(incomeList), HttpStatus.OK);
-    }
-    
     @RequestMapping(value = "/update/{id}", method = RequestMethod.POST)
     public ResponseEntity<String> updateIncome(@PathVariable("id") Long id, @RequestBody @Valid Income income) {
     
+        String loggedUsername = ControllerUtil.getCurrentLoggedUsername();
         Income oldIncome = incomeDao.findOne(id);
-        if (oldIncome == null) {
+        if (oldIncome == null || !(loggedUsername.equals(oldIncome.getUser().getUsername()))) {
             return new ResponseEntity<String>("Income not found", HttpStatus.NOT_FOUND);
         }
         income.setId(id);
+        income.setUser(oldIncome.getUser());
         incomeDao.saveAndFlush(income);
         return new ResponseEntity<String>("Income updated", HttpStatus.NO_CONTENT);
     }
@@ -92,6 +101,11 @@ public class IncomeController {
     public ResponseEntity<String> deleteIncome(@PathVariable("id") Long id) {
     
         try {
+            String loggedUsername = ControllerUtil.getCurrentLoggedUsername();
+            Income incomeToBeDeleted = incomeDao.findOne(id);
+            if (incomeToBeDeleted == null || !(loggedUsername.equals(incomeToBeDeleted.getUser().getUsername()))) {
+                throw new EmptyResultDataAccessException("Income not found", 1);
+            }
             incomeDao.delete(id);
         } catch (EmptyResultDataAccessException emptyResultDataAccessException) {
             log.info(emptyResultDataAccessException.getMessage());
@@ -100,10 +114,12 @@ public class IncomeController {
         return new ResponseEntity<String>("Income deleted", HttpStatus.NO_CONTENT);
     }
     
-    @RequestMapping(value = "/deleteAll/", method = RequestMethod.DELETE)
+    @RequestMapping(value = "/delete_all", method = RequestMethod.DELETE)
     public ResponseEntity<String> deleteAll() {
     
-        incomeDao.deleteAll();
+        String loggedUsername = ControllerUtil.getCurrentLoggedUsername();
+        incomeDao.deleteAllByUsername(loggedUsername);
+        incomeDao.flush();
         return new ResponseEntity<String>("Incomes deleted", HttpStatus.NO_CONTENT);
     }
